requirements.txt
dash==2.10.2
dash_bootstrap_components==0.13.1
dash_html_components==2.0.0
folium==0.14.0
geopy==2.3.0
python-dotenv==1.0.0
requests==2.31.0
